# public/swagger-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Capstone - API - Documentation",
    "description": "API for user interface development",
    "version": "1.1",
    "contact": {
      "name": "mangox API Support",
      "email": "support@mangox.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/front/detail/{slug}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "slug",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_front_detail__slug_",
        "tags": [
          "Front"
        ]
      }
    },
    "/api/v1/front/detail/{postTypeCollectionName}/{SEOPathSlug}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "postTypeCollectionName",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "SEOPathSlug",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_front_detail__postTypeCollectionName___SEOPathSlug_",
        "tags": [
          "Front"
        ]
      }
    },
    "/api/v1/front/{*}": {
      "put": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "*",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "put__api_v1_front____",
        "tags": [
          "Front"
        ]
      },
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "*",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "get__api_v1_front____",
        "tags": [
          "Front"
        ]
      },
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "*",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "post__api_v1_front____",
        "tags": [
          "Front"
        ]
      },
      "patch": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "*",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "patch__api_v1_front____",
        "tags": [
          "Front"
        ]
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "*",
            "required": true
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "default": "68a6a4a4e731734fad2cff97"
            },
            "description": "Tenant ID for multi-tenant support"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "operationId": "delete__api_v1_front____",
        "tags": [
          "Front"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://mangox-api.mangoads.com.vn",
      "description": "Development Server"
    },
    {
      "url": "http://localhost:5555",
      "description": "Development Server"
    }
  ],
...
  "security": [
    {},
    {
      "bearerAuth": []
    }
  ]
}